Crate nybbles

nybbles

Efficient nibble-sized (4-bit) byte sequence data structure.

Structs

Nibbles

Structure representing a sequence of nibbles.

Functions

common_prefix_length

Returns the length of the common prefix between the two slices.

get_byte

Gets the byte at the given index by combining two consecutive nibbles.

get_byte_unchecked⚠

Gets the byte at the given index by combining two consecutive nibbles.

pack_to

Packs the nibbles into the given slice.

pack_to_unchecked⚠

Packs the nibbles into the given slice without checking its length.

smallvec_with⚠

Initializes a smallvec with the given length and a closure that initializes the buffer.